Refactor resoures with TagResourceReference, add Gen3 resource cache, refactor gen3 bitmap resource definition, add automatic resource size calculation and extraction

+++ TagTool/Serialization/Gen3SerializationContext.cs
@@ -55,7 +55,7 @@ namespace TagTool.Serialization
 
         public CachedTagInstance GetTagByIndex(int index)
         {
-            var tag = GameCache.TagCache.GetTagByID(index);
+            var tag = GameCache.TagCache.GetTagByID((uint)index);
 
             var group = (tag != null) ? tag.Group : TagGroup.None;
